Create project1.s
my student id is 02968946, and 02968946 % 11 gives 2. So, the  Δ  is 'R' and β is 'r' for me. In the given line of codes, we allocate 10 character spaces for the input in the data section and in the text section of the code, we take input from the user and store it in the register $a0.