Update project1.s
we create three labels uppercase, lowercase and numbers, and based on the character we read we would jump to the suitable labels. Also since the input for me is restricted to only alphabets and numbers, we ignore anything that's not an alphabet or a number which is equivalent to adding 0.  Also we move the piece of code before the iteration label, instead of having to jump to iteration each time.