Update project1.s
For uppercase letters, we compared if the ascii value of the given character is equal or more than 83 ('S'). If so, we jump back to the label iterations again as the value to be added will be 0. Since the value for 'A' is 10 and 'R' is 27, we get it by subtracting the ascii value by 55. We now check if the value of register $t3 (the index) is even or odd. We use the instruction andi that gives 1 if odd and 0 if even and store the value in a new register $t5. We check if the current index is even or odd because we will add even and odd indices seperately and subtract the even from the odd.