Update project1.s
we previously checked if the value of the current index is even or odd and stored it in the register $t5 (0 for even and 1 for odd). In the next line, we us the instruction bnez to jump to the label odd whenever the value of $t1 is not 0 (i.e. 1).  We record the sum of the register $t4 to a new register ($t6 if the index is even and $t7 if the index is odd). We do this so we can subtract the values at the end of the program.