Update project1.s
Similarly, after we checked if the character is lowercase and jumped to the label lowercase, we again check to see if the ascii value for the character is equal or greater than the ascii code for 's' (115). If so, we jump back to iteration. Otherwise, we subtract 87 from it to get the decimal equivalent of the character. Again, we check if the index is odd or even, and store the value in $t5 (0 for even and 1 for odd). if it's odd we jump to the label odd. If not, we add the decimal equivalent value stored in $t4 to the register $t6. If it's odd, we jump to the label odd which add the value to the register $t7.