Update project1.s
We, similarly, after jumping to the numbers again check if the characters are valid numbers (0 to 9). If not, we jump back to iteration. Otherwise, we subtract 48 from the ascii value to get the decimal value and store it in $t4. We then check if the current index is even or odd. If it is odd, the program jumps to the label odd.