Update project1.s
We check if the previously stored value in $t5 is odd or even (0 for even and 1 for odd). If odd, we jump to the label odd where we add the decimal equivalent of the character stored in $t4 to the register $t7. if even, we add the value to the register $t6.