Update project1.s
In the last commit, I forgot to mention that in the end, we subtracted the value in register $t6 from the value in $t7. This gives us the sum of all the elements in even indices subtracted from the sum of the elements in odd indices.  We store the sum in the register $t7. Then we move the value to $a0 to print the value and we finally make an exit to the program.